Move utils and components to core module
